style: run forge fmt

+++ contracts-periphery/src/interface/ISwapRouter.sol
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: GPL-2.0-or-later
 pragma solidity ^0.8.13;
 
-import 'src/interface/IUniswapV3SwapCallback.sol';
+import "src/interface/IUniswapV3SwapCallback.sol";
 
 /// @title Router token swapping functionality
 /// @notice Functions for swapping tokens via Uniswap V3
@@ -9,15 +9,22 @@ interface ISwapRouter is IUniswapV3SwapCallback {
   struct ExactInputParams {
     bytes path;
     address recipient;
-    uint256 amountIn;
-    uint256 amountOutMinimum;
+    uint amountIn;
+    uint amountOutMinimum;
   }
-  function exactInput(ExactInputParams calldata params) external payable returns (uint256 amountOut);
-  function multicall(bytes[] calldata data) external payable returns (bytes[] memory results);
+
+  function exactInput(ExactInputParams calldata params)
+    external
+    payable
+    returns (uint amountOut);
+  function multicall(bytes[] calldata data)
+    external
+    payable
+    returns (bytes[] memory results);
   function unwrapWETH9WithFee(
-    uint256 amountMinimum,
+    uint amountMinimum,
     address recipient,
-    uint256 feeBips,
+    uint feeBips,
     address feeRecipient
   ) external payable;
 }
